upd_db/requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
dnspython==2.0.0
finnhub-python==2.0.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.1
pandas==1.0.5
pylint==2.5.3
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
